python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2

# Google API packages
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0

# Supporting packages (optional but recommended)
python-dotenv>=0.19.0  # For environment variables
pyOpenSSL>=20.0.0      # For secure connections
psycopg2-binary>=2.9.9  # PostgreSQL database adapter
SQLAlchemy>=2.0.0